Add ActionMenuIcon component
